include/iprt/asm.h: The bitmap needs to be aligned on a 32-bit address only and not pointer size aligned
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@99563 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/asm.h
@@ -6819,7 +6819,7 @@ DECLINLINE(bool) ASMBitTest(const volatile void RT_FAR *pvBitmap, int32_t iBit) 
 
 # else
     int32_t offBitmap = iBit / 32;
-    AssertRelease(!((uintptr_t)pvBitmap & (sizeof(uintptr_t) - 1))); /* This assumes a power of two uintptr_t size. */
+    AssertRelease(!((uintptr_t)pvBitmap & (sizeof(uint32_t) - 1)));
     rc.u32 = RT_LE2H_U32(ASMAtomicUoReadU32(&((uint32_t volatile *)pvBitmap)[offBitmap])) >> (iBit & 31);
     rc.u32 &= 1;
 # endif
